Delete po napustanju, relociranje fileova, poboljsanja
